Implements digging, kind of.
Factors player movement out of the main game loop.
Adds dig strength as a property of an entity, later to be factored out
into an item.
Implements a basic loop so that entities may be updated only after the
player has moved, and so map data is only shuttled to VRAM if something
has updated.